requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Werkzeug==2.2.3
tensorflow==2.14.0
Pillow==10.0.0
numpy==1.26.0
gunicorn==20.1.0
twilio==8.4.1
gTTS==2.4.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
